separate def
