+++ crates/goose-providers/src/lib.rs
@@ -12,7 +12,6 @@ pub(crate) mod mcp_utils;
 pub mod model;
 pub mod openai;
 pub mod openai_compatible;
-pub mod openai_def;
 pub mod permission;
 pub mod request_log;
 pub mod retry;
