+++ crates/goose/src/config/declarative_providers.rs
@@ -6,7 +6,7 @@ use crate::providers::huggingface::HuggingFaceProvider;
 use crate::providers::huggingface_auth;
 use crate::providers::inventory::declarative_inventory_identity;
 use crate::providers::ollama::OllamaProvider;
-use crate::providers::openai::OpenAiProvider;
+use crate::providers::openai_def::OpenAiProviderDef;
 use anyhow::Result;
 use include_dir::{include_dir, Dir};
 use once_cell::sync::Lazy;
@@ -604,14 +604,14 @@ pub fn register_declarative_provider(
                         },
                     );
             } else {
-                registry.register_with_name::<OpenAiProvider, _, _>(
+                registry.register_with_name::<OpenAiProviderDef, _, _>(
                     &config,
                     provider_type,
                     config.dynamic_models.unwrap_or(false),
                     move |model, tls_config| {
                         let mut cfg = captured.clone();
                         resolve_config(&mut cfg)?;
-                        OpenAiProvider::from_custom_config(model, cfg, tls_config)
+                        crate::providers::openai_def::from_custom_config(model, cfg, tls_config)
                     },
                     move || {
                         let mut cfg = identity_config.clone();
