+++ crates/goose/src/providers/init.rs
@@ -32,7 +32,6 @@ use super::{
     litellm::LiteLLMProvider,
     nanogpt::NanoGptProvider,
     ollama::OllamaProvider,
-    openai::OpenAiProvider,
     openrouter::OpenRouterProvider,
     pi_acp::PiAcpProvider,
     provider_registry::ProviderRegistry,
@@ -43,6 +42,7 @@ use super::{
 };
 use crate::config::ExtensionConfig;
 use crate::providers::base::ProviderType;
+use crate::providers::openai_def::OpenAiProviderDef;
 use crate::{
     config::declarative_providers::register_declarative_providers,
     providers::provider_registry::ProviderEntry,
@@ -120,7 +120,7 @@ async fn init_registry() -> RwLock<ProviderRegistry> {
             true,
             Some(registrations::ollama_inventory()),
         );
-        registry.register_with_inventory::<OpenAiProvider>(
+        registry.register_with_inventory::<OpenAiProviderDef>(
             true,
             Some(registrations::openai_inventory()),
         );
