+++ crates/goose/src/providers/openai_def.rs
@@ -1,19 +1,41 @@
 use anyhow::Result;
+use futures::future::BoxFuture;
+use goose_providers::base::ProviderDescriptor;
 use std::collections::HashMap;
 
-use crate::api_client::{ApiClient, AuthMethod};
-use crate::base::DEFAULT_PROVIDER_TIMEOUT_SECS;
 use crate::config::declarative_providers::DeclarativeProviderConfig;
-use crate::model::ModelConfig;
-use crate::openai::{
-    ensure_url_scheme, parse_custom_headers, OpenAiProvider, OpenAiProviderBuilder, ParsedBaseUrl,
-    OPEN_AI_DEFAULT_BASE_PATH, OPEN_AI_DEFAULT_FAST_MODEL, OPEN_AI_PROVIDER_NAME,
-    OPEN_AI_VERSIONLESS_BASE_PATH,
+use crate::providers::base::{ProviderDef, DEFAULT_PROVIDER_TIMEOUT_SECS};
+use goose_providers::api_client::{ApiClient, AuthMethod};
+use goose_providers::model::ModelConfig;
+use goose_providers::openai::{
+    ensure_url_scheme, parse_custom_headers, parse_openai_base_url, OpenAiProvider,
+    OpenAiProviderBuilder, OPEN_AI_DEFAULT_BASE_PATH, OPEN_AI_DEFAULT_FAST_MODEL,
+    OPEN_AI_PROVIDER_NAME, OPEN_AI_VERSIONLESS_BASE_PATH,
 };
+
+pub struct OpenAiProviderDef;
+
+impl ProviderDescriptor for OpenAiProviderDef {
+    fn metadata() -> goose_providers::base::ProviderMetadata {
+        OpenAiProvider::metadata()
+    }
+}
+
+impl ProviderDef for OpenAiProviderDef {
+    type Provider = OpenAiProvider;
+
+    fn from_env(
+        model: ModelConfig,
+        _extensions: Vec<crate::config::ExtensionConfig>,
+        tls_config: Option<crate::providers::api_client::TlsConfig>,
+    ) -> BoxFuture<'static, Result<Self::Provider>> {
+        Box::pin(from_env(model, tls_config))
+    }
+}
 
 pub async fn from_env(
     model: ModelConfig,
-    tls_config: Option<crate::api_client::TlsConfig>,
+    tls_config: Option<goose_providers::api_client::TlsConfig>,
 ) -> Result<OpenAiProvider> {
     let config = crate::config::Config::global();
 
@@ -48,7 +70,7 @@ pub async fn from_env(
         .map(|s| s.trim().to_string())
         .filter(|s| !s.is_empty())
     {
-        OpenAiProvider::parse_base_url(&raw_url)?
+        parse_base_url(&raw_url)?
     } else {
         let h: String = config
             .get_param("OPENAI_HOST")
@@ -215,7 +237,7 @@ pub fn resolve_api_key(
 pub fn from_custom_config(
     model: ModelConfig,
     config: DeclarativeProviderConfig,
-    tls_config: Option<crate::api_client::TlsConfig>,
+    tls_config: Option<goose_providers::api_client::TlsConfig>,
 ) -> Result<OpenAiProvider> {
     let custom_models = if !config.models.is_empty() {
         Some(
@@ -257,7 +279,7 @@ pub fn from_custom_config(
     let base_path = if let Some(ref explicit_path) = config.base_path {
         explicit_path.trim_start_matches('/').to_string()
     } else {
-        OpenAiProvider::derive_base_path(url.path())
+        derive_base_path(url.path())
     };
 
     let timeout_secs = config
@@ -301,4 +323,166 @@ pub fn from_custom_config(
         .skip_canonical_filtering(config.skip_canonical_filtering)
         .preserve_thinking_context(config.preserves_thinking)
         .build())
+}
+
+/// Components extracted from an `OPENAI_BASE_URL` value.
+struct ParsedBaseUrl {
+    /// The host (scheme + authority + any path prefix before `/v1`).
+    pub(crate) host: String,
+    /// Query parameters to forward on every request.
+    pub(crate) query_params: Vec<(String, String)>,
+    /// Whether the URL path ended with `/v1`.
+    pub(crate) has_v1: bool,
+    /// `true` when the host was derived from `OPENAI_BASE_URL`.
+    /// Controls whether `OPENAI_BASE_PATH` is read from env only
+    /// (to avoid persisted desktop defaults shadowing URL-derived paths)
+    /// or from config too (to honour Docker Model Runner setups).
+    pub(crate) from_base_url: bool,
+}
+
+fn parse_base_url(raw_url: &str) -> Result<ParsedBaseUrl> {
+    let (host, query_params, has_v1) = parse_openai_base_url(raw_url)?;
+    Ok(ParsedBaseUrl {
+        host,
+        query_params,
+        has_v1,
+        from_base_url: true,
+    })
+}
+
+fn derive_base_path(url_path: &str) -> String {
+    let stripped = url_path.trim_start_matches('/');
+    let normalized = stripped.trim_end_matches('/');
+    if normalized.is_empty() {
+        "v1/chat/completions".to_string()
+    } else if normalized.ends_with("chat/completions") {
+        stripped.to_string()
+    } else if ends_with_version_segment(normalized) {
+        format!("{}/chat/completions", normalized)
+    } else {
+        format!("{}/v1/chat/completions", normalized)
+    }
+}
+
+fn ends_with_version_segment(path: &str) -> bool {
+    let last = path.rsplit('/').next().unwrap_or(path);
+    last.strip_prefix('v')
+        .is_some_and(|rest| !rest.is_empty() && rest.bytes().all(|b| b.is_ascii_digit()))
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn parse_base_url_strips_v1_from_standard_openai_url() {
+        let r = parse_base_url("https://api.openai.com/v1").unwrap();
+        assert_eq!(r.host, "https://api.openai.com");
+        assert!(r.query_params.is_empty());
+        assert!(r.has_v1);
+    }
+
+    #[test]
+    fn parse_base_url_preserves_prefix_before_v1() {
+        let r = parse_base_url("https://gateway.example.com/openai/v1").unwrap();
+        assert_eq!(r.host, "https://gateway.example.com/openai");
+        assert!(r.has_v1);
+    }
+
+    #[test]
+    fn parse_base_url_handles_no_path() {
+        let r = parse_base_url("https://api.openai.com").unwrap();
+        assert_eq!(r.host, "https://api.openai.com");
+        assert!(r.has_v1);
+    }
+
+    #[test]
+    fn parse_base_url_handles_trailing_slash() {
+        let r = parse_base_url("https://api.openai.com/v1/").unwrap();
+        assert_eq!(r.host, "https://api.openai.com");
+        assert!(r.has_v1);
+    }
+
+    #[test]
+    fn parse_base_url_preserves_port() {
+        let r = parse_base_url("https://localhost:8080/v1").unwrap();
+        assert_eq!(r.host, "https://localhost:8080");
+        assert!(r.has_v1);
+    }
+
+    #[test]
+    fn parse_base_url_preserves_non_v1_path() {
+        let r = parse_base_url("https://example.com/custom/api").unwrap();
+        assert_eq!(r.host, "https://example.com/custom/api");
+        assert!(!r.has_v1);
+    }
+
+    #[test]
+    fn derive_base_path_not_removing_api_path() {
+        let r = derive_base_path("https://opencode.ai/zen/go");
+        assert_eq!(r, "https://opencode.ai/zen/go/v1/chat/completions");
+    }
+
+    #[test]
+    fn derive_base_path_should_support_v1() {
+        let r = derive_base_path("https://opencode.ai/zen/go/v1");
+        assert_eq!(r, "https://opencode.ai/zen/go/v1/chat/completions");
+    }
+
+    #[test]
+    fn derive_base_path_should_support_no_base_path() {
+        let r = derive_base_path("https://opencode.ai/");
+        assert_eq!(r, "https://opencode.ai/v1/chat/completions");
+    }
+
+    #[test]
+    fn derive_base_path_preserves_non_v1_version_prefix() {
+        // Zhipu's default base_url is https://open.bigmodel.cn/api/paas/v4 and
+        // from_custom_config passes url.path() ("/api/paas/v4") here. The
+        // existing /api/paas/v4 version must not gain an extra /v1 segment.
+        let r = derive_base_path("/api/paas/v4");
+        assert_eq!(r, "api/paas/v4/chat/completions");
+    }
+
+    #[test]
+    fn derive_base_path_does_not_treat_v_word_as_version() {
+        let r = derive_base_path("/api/voice");
+        assert_eq!(r, "api/voice/v1/chat/completions");
+    }
+
+    #[test]
+    fn parse_base_url_preserves_query_params() {
+        let r = parse_base_url("https://gw.example.com/v1?api-version=2024-02-01").unwrap();
+        assert_eq!(r.host, "https://gw.example.com");
+        assert_eq!(
+            r.query_params,
+            vec![("api-version".to_string(), "2024-02-01".to_string())]
+        );
+        assert!(r.has_v1);
+    }
+
+    #[test]
+    fn parse_base_url_preserves_multiple_query_params() {
+        let r = parse_base_url("https://example.com/v1?key=val&foo=bar").unwrap();
+        assert_eq!(r.query_params.len(), 2);
+        assert_eq!(r.query_params[0], ("key".to_string(), "val".to_string()));
+        assert_eq!(r.query_params[1], ("foo".to_string(), "bar".to_string()));
+    }
+
+    #[test]
+    fn parse_base_url_preserves_credentials() {
+        let r = parse_base_url("https://user:pass@gateway.example.com/v1").unwrap();
+        assert_eq!(r.host, "https://user:pass@gateway.example.com");
+        assert!(r.has_v1);
+    }
+
+    #[test]
+    fn parse_base_url_rejects_empty_string() {
+        assert!(parse_base_url("").is_err());
+    }
+
+    #[test]
+    fn parse_base_url_rejects_whitespace_only() {
+        assert!(parse_base_url("  ").is_err());
+    }
 }
